requirements.txt
beautifulsoup4==4.12.0
Flask==2.3.2
ImageHash==4.3.1
numpy
Pillow==9.5.0
Requests==2.31.0
Scrapy==2.9.0
scikit-image==0.21.0
gunicorn
opencv-python